updates for MacOS

+++ install.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-files=$(find . -maxdepth 1 -type f -not -name '*.sh' -printf "%f\n")
+files=$(find . -maxdepth 1 -type f -not -name '*.sh' -exec basename {} \;)
 echo "Installing dotfiles:" $files
 for f in $files; do
   rm "${HOME}/.${f}" 2> /dev/null
@@ -8,11 +8,13 @@ for f in $files; do
 done
 
 
-echo "Installing i3 config"
-rm ~/.config/i3/config 2> /dev/null
+if [ "$(uname -s)" == "Linux" ]; then
+  echo "Installing i3 config"
+  rm ~/.config/i3/config 2> /dev/null
 
-if [ ! -d ${HOME}/.config/i3 ]; then
-  mkdir -p ${HOME}/.config/i3
+  if [ ! -d ${HOME}/.config/i3 ]; then
+    mkdir -p ${HOME}/.config/i3
+  fi
+
+  ln -s ${PWD}/i3/config ${HOME}/.config/i3/config
 fi
-
-ln -s ${PWD}/i3/config ${HOME}/.config/i3/config
